+++ engine/flink/engine/src/main/scala/pl/touk/nussknacker/engine/process/exception/FlinkExceptionHandler.scala
@@ -1,13 +1,14 @@
-package pl.touk.nussknacker.engine.flink.api.exception
+package pl.touk.nussknacker.engine.process.exception
 
 import com.typesafe.config.Config
 import net.ceedubs.ficus.Ficus.{booleanValueReader, optionValueReader, stringValueReader, toFicusConfig}
 import org.apache.flink.api.common.restartstrategy.RestartStrategies
 import pl.touk.nussknacker.engine.api.exception.NuExceptionInfo
-import pl.touk.nussknacker.engine.api.{Context, Lifecycle, MetaData, ProcessListener}
 import pl.touk.nussknacker.engine.api.process.ProcessObjectDependencies
 import pl.touk.nussknacker.engine.api.runtimecontext.EngineRuntimeContext
-import pl.touk.nussknacker.engine.flink.api.exception.FlinkExceptionHandler._
+import pl.touk.nussknacker.engine.api.{Context, MetaData, ProcessListener}
+import pl.touk.nussknacker.engine.flink.api.exception.{ExceptionHandler, FlinkEspExceptionConsumer, FlinkEspExceptionConsumerProvider, RateMeterExceptionConsumer}
+import pl.touk.nussknacker.engine.process.exception.FlinkExceptionHandler.{exceptionHandlerConfigPath, typeConfigPath, withRateMeterPath}
 import pl.touk.nussknacker.engine.util.exception.WithExceptionExtractor
 import pl.touk.nussknacker.engine.util.loader.ScalaServiceLoader
