+++ engine/flink/engine/src/test/scala/pl/touk/nussknacker/engine/process/exception/FlinkExceptionHandlerSpec.scala
@@ -1,14 +1,14 @@
-package pl.touk.nussknacker.engine.flink.api.exception
+package pl.touk.nussknacker.engine.process.exception
 
-import com.typesafe.config.ConfigFactory
+import com.typesafe.config.{Config, ConfigFactory}
 import org.apache.flink.api.common.restartstrategy.RestartStrategies
+import org.scalatest.{FunSuite, Matchers}
 import pl.touk.nussknacker.engine.api.exception.{NonTransientException, NuExceptionInfo}
 import pl.touk.nussknacker.engine.api.process.ProcessObjectDependencies
 import pl.touk.nussknacker.engine.api.{Context, MetaData, StreamMetaData}
+import pl.touk.nussknacker.engine.flink.api.exception.{FlinkEspExceptionConsumer, FlinkEspExceptionConsumerProvider}
 import pl.touk.nussknacker.engine.util.namespaces.DefaultNamespacedObjectNaming
 import pl.touk.nussknacker.test.ClassLoaderWithServices
-import com.typesafe.config.Config
-import org.scalatest.{FunSuite, Matchers}
 
 import scala.jdk.CollectionConverters.mapAsJavaMapConverter
