+++ engine/flink/engine/src/test/scala/pl/touk/nussknacker/engine/process/exception/RestartStrategyFromConfigurationSpec.scala
@@ -1,4 +1,4 @@
-package pl.touk.nussknacker.engine.flink.api.exception
+package pl.touk.nussknacker.engine.process.exception
 
 import com.typesafe.config.ConfigFactory
 import org.apache.flink.api.common.restartstrategy.RestartStrategies
